backend/requirements.txt
# Prometheus Backend - Form Analysis Service with MediaPipe, VBT, and YOLO
# Python 3.10-3.12 required

# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
aiofiles==23.2.1
pydantic==2.5.0

# Pose estimation with MediaPipe
mediapipe>=0.10.8

# VBT calculations (signal processing)
scipy>=1.9.0

# YOLO for barbell detection (VBT)
ultralytics>=8.0.0

# Video processing
opencv-python-headless==4.8.1.78  # Headless version for server deployment
numpy==1.24.3

# Supabase client
supabase>=2.0.0

# AI Coach - OpenAI API
openai>=1.0.0

# HTTP client for downloading videos and Revolut API
httpx>=0.25.0

# Production deployment
gunicorn==21.2.0
python-dotenv==1.0.0
